--- model/elements/src/main/kotlin/de/gleex/pltcmd/model/elements/Unit.kt
@@ -1,10 +1,0 @@
-package de.gleex.pltcmd.model.elements
-
-/**
- * A military unit is the smallest part of an [Element].
- */
-interface Unit {
-
-    fun isOfType(type: UnitType): Boolean
-
-}
