--- model/elements/src/main/kotlin/de/gleex/pltcmd/model/elements/UnitType.kt
@@ -1,7 +1,0 @@
-package de.gleex.pltcmd.model.elements
-
-enum class UnitType {
-    Car,
-    Helicopter,
-    Soldier
-}
