--- model/elements/src/main/kotlin/de/gleex/pltcmd/model/elements/reworked/dsl/elementsDSL.kt
@@ -1,2 +1,0 @@
-package de.gleex.pltcmd.model.elements.reworked.dsl
-
